Split storage into sqlstorage and a part-implemented datomstore.js.
